services/rag/app/requirements.txt
fastapi
uvicorn
requests
qdrant-client
sentence-transformers
pypdf
python-docx
beautifulsoup4
chardet